+++ apyfal/__init__.py
@@ -90,6 +90,7 @@ class Accelerator(object):
                 host_ip = self._host.url
             except _exc.HostException:
                 host_ip = None
+
         else:
             # Use local host
             self._host = None
@@ -98,10 +99,18 @@ class Accelerator(object):
             client_type = 'REST' if (host_ip and host_type is None) else None
 
         # Create AcceleratorClient object
+
+        try:
+            # Get SSL certificate to verify
+            ssl_cert_crt = self._host.ssl_cert_crt
+        except AttributeError:
+            ssl_cert_crt = None
+
         self._client = _clt.AcceleratorClient(
             accelerator=accelerator, client_type=client_type,
             accelize_client_id=accelize_client_id, host_ip=host_ip,
-            accelize_secret_id=accelize_secret_id, config=config)
+            accelize_secret_id=accelize_secret_id, config=config,
+            ssl_cert_crt=ssl_cert_crt)
 
     def __enter__(self):
         return self